AI_Mentor/requirements.txt
Flask
langchain
openai
faiss-cpu
chromadb
beautifulsoup4
requests
pandas
ujson
python-dotenv
gunicorn
